Add files via upload
This dashboard was develop to answer two questions that a games distribution company has: 
How many clients subscribe their games in a automatic yearly renowal ? And how big is the annual income with it? 

The first step of this project was to bild a database with information like: game price, kind of renowal, subscription type, etc...
Made this, we started calculating the number of people with subscription and the kinds of renowals and plans they rather.
After all this done, we follow to the dashboard construction. In this fase, is very importat that you dont use to many collors, and it must be in a legible form, big enough to see it.  However, the master key to deliver a successful dashboard is to answer the the companys's questions in the clearlyest way possible, in a way which the only thinking time for them to spend is how thier strategy follows from there!!